// libs/shared/sofy/src/lib/models/sofy-product.ts
import { OptionInterface } from '../interfaces/option.interface';
import { SofyAddress } from './sofy-address';

export class SofyProduct {
  omapiid?: string;
  options: OptionInterface[];
  isnewline: boolean;
  installationaddress: SofyAddress;
  move: boolean;
  needscableinstallation: boolean;
  installtype: string;
}