// libs/shared/sofy/src/lib/models/request/sofy-request-product.ts
import { SofyRequestAddress } from './sofy-request-address';
import { SofyRequestStreamingService } from './sofy-request-streaming-service';

export interface SofyRequestProduct {
  correlationid?: string;
  omapiid: string;
  installationaddress: SofyRequestAddress;
  options: string[];
  streamingservices?: SofyRequestStreamingService[];
  optinproducts: string[];
  installtype?: string;
  amountoflimitedsims?: number;
}